fixed bug

+++ app/src/main/java/com/bookshelf/app/dashboard/data/models/Year.kt
@@ -1,3 +1,3 @@
 package com.bookshelf.app.dashboard.data.models
 
-data class Year(var year: Long? = null, var isActive: Boolean = false)
+data class Year(var year: Int? = null, var isActive: Boolean = false)
